requirements.txt
# Bot de Trading NASDAQ 100
# Pour Railway

alpaca-trade-api==3.2.0
ta==0.11.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
schedule==1.2.1
pytz==2024.1
requests==2.31.0
